+++ packages/forms/src/FormFieldDatePicker.tsx
@@ -19,7 +19,6 @@ export const FormFieldDatePicker = <
   TFieldValues extends FieldValues = FieldValues,
   TName extends FieldPath<TFieldValues> = FieldPath<TFieldValues>,
 >({
-  disabled,
   ...props
 }: FormFieldDatePickerProps<TFieldValues, TName>) => {
   const {
@@ -34,6 +33,7 @@ export const FormFieldDatePicker = <
     defaultValue,
     presets,
     unsavedChangesGuard,
+    disabled,
     ...datePickerProps
   } = props;
 
@@ -54,6 +54,7 @@ export const FormFieldDatePicker = <
         return (
           <DatePicker
             {...datePickerProps}
+            disabled={disabled}
             value={field.value as DateRange | undefined}
             onChange={(range: DateRange | undefined) => {
               field.onChange(range);