+++ challenge/packages/app-vite-with-graphql-api/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.1](https://github.com/ttoss/ttoss/compare/@ttoss-challenge/app-vite-with-graphql-api@0.3.0...@ttoss-challenge/app-vite-with-graphql-api@0.3.1) (2026-02-23)
+
+**Note:** Version bump only for package @ttoss-challenge/app-vite-with-graphql-api
+
 # [0.3.0](https://github.com/ttoss/ttoss/compare/@ttoss-challenge/app-vite-with-graphql-api@0.2.16...@ttoss-challenge/app-vite-with-graphql-api@0.3.0) (2026-02-15)
 
 ### Features