+++ packages/forms/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.41.1](https://github.com/ttoss/ttoss/compare/@ttoss/forms@0.41.0...@ttoss/forms@0.41.1) (2026-02-23)
+
+**Note:** Version bump only for package @ttoss/forms
+
 # [0.41.0](https://github.com/ttoss/ttoss/compare/@ttoss/forms@0.40.2...@ttoss/forms@0.41.0) (2026-02-15)
 
 ### Features