+++ packages/react-dashboard/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.1](https://github.com/ttoss/ttoss/compare/@ttoss/react-dashboard@0.5.0...@ttoss/react-dashboard@0.5.1) (2026-02-23)
+
+**Note:** Version bump only for package @ttoss/react-dashboard
+
 # [0.5.0](https://github.com/ttoss/ttoss/compare/@ttoss/react-dashboard@0.4.0...@ttoss/react-dashboard@0.5.0) (2026-02-15)
 
 ### Features